GeneralRedirect to use snapshots instead of deprecated useRecoilInterface
Summary:
Use mutable snapshots instead of useRecoilInterface which is deprecated. This fixes latent bugs stemming from the expectation that we can read our writes. Reading our writes is accomplished using snapshots.

In this diff we attempt to not change the logic at all, but because the snapshot interface is async a bunch of code has to move to async.

* Removes a setTimeout in setting `selected_ds`, which is the date selection for CV Home. Based on manual testing this setTimeout appears to be unnecessary. My guess is it was a vestige from when you couldn't set both Recoil and NavigationState state in the same event.

Reviewed By: drarmstr

Differential Revision: D22353594

fbshipit-source-id: 59d8a01d460c04c0e5b997846f6e6e64790efab2