Recoil transactions should create a new TreeState version (#463)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/463

Recoil tracks state versions and updates each version with a batch commit.  The version is important for tracking async resolutions and concurrent mode.  Going to a previous `Snapshot` with `useGotoRecoilSnapshot()` would revert to the previous `TreeState` versions, which is good.  However, going to a mutated snapshot for atomic transactions did not update the `TreeState` version.  Also, going to a previous snapshot would then simply increment the version with subsequent state changes and conflict with that other previous version.

This diff fixes that by using a new version for each batch commit, `Snapshot` mutation, or fresh `Snapshot` construction.

The state version can be exposed as a `Snapshot` ID for equality checks.  This allows dev tools to check if we are going to a previous Snapshot with `useGotoRecoilSnapshot()`. (cc maxijb)

Reviewed By: davidmccabe

Differential Revision: D22447073

fbshipit-source-id: cfa3e5c8f3bc6bef8c8c7592a7adf33394715cf3