Protect from impure state updater functions
Summary:
Throw an error with an explanatory message is case an impure function is passed as a state updater which causes another atom update within the execution of that function. This condition happened to work before but recent changes have meant that it doesn't work. It has always been contrary to the contract of the API which expects a pure function.

This problem was encountered in product code which was fixed in D23670967.

Reviewed By: drarmstr

Differential Revision: D23687449

fbshipit-source-id: 4413f673b0d7dcf7793d2244307eb587d7ca303a