Adding Component Debug Names
Summary:
Currently, there is no way to determine at runtime, which React Components are subscribed to which Recoil Nodes. The issue is that React does not currently have an API for finding what the calling React Component is for a hook (https://fb.workplace.com/groups/233834187293137/permalink/304610230215532/).

This diff hacks through this by using a couple of assumptions:
1. Hooks are only ever called by React Components, or other Hooks
2. Hook functions always start with "use" and React components never start with "use"

We can therefore look at the current stack within the hook at runtime, and walk down until we find the first entry which does not start with "use". If the assumptions hold then this must be the calling React component.

This is fine in dev, however in prod things become less useful due to minification, and also the assumptions may no longer hold. This feature has therefore been guarded by the __DEV__ flag global.

Reviewed By: drarmstr, csantos42

Differential Revision: D21957843

fbshipit-source-id: 00363e5ada437c0051984e8eec615322c55539a2