O(edges) selector traversal
Summary:
Cache selector results until the selector has a dependency that changes. This prevents needing to evaluating the selector's deps in order to compute a cache key to look up in the map.

Depending on the shape of the selector graph, this can yield huge benefits.

CV Setup page had a 30-40 second block of solid CPU that is now about 1 second. This is because selector edge traversal went from around 5100 to around 470.

This doesn't yet apply to async selectors in a loading state. I think this would be possible but it's more complex and I need to think about it more.

Reviewed By: drarmstr

Differential Revision: D23945882

fbshipit-source-id: de81c4b11829b0cd92f2f3588eb278140a73051a