Add error reporting for attempted async selector set (#777)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/777

Recoil does not support async selector sets.  However, as it does support async gets many users expect async sets.  Flow reports a static error if an `async` function is used, TypeScript should as well.  Though users may not have or notice this error.  This diff throws an error if callbacks in the `set` are attempted to be used asynchronously.  However, those errors may not be caughts and be lost.  So, we also detect if an `async` function was used by checking if the user's set returns a `Promise` and throws an error for that as well.

Reviewed By: davidmccabe

Differential Revision: D25250686

fbshipit-source-id: 93c529f09734142cdd04ea0eda7097f814b120cb