add support for user thrown promises in selector_NEW
Summary:
The new selector implementation currently does not support user-thrown promises, meaning the user cannot throw a loadable promise (or an entirely custom promise created by the user to control exactly when the selector should re-execute).

Under normal conditions the selector checks a "dep cache" (a cache containing dep values for the current execution) when re-executing as a performance optimization to avoid having to re-check the cache of all selectors in the dependency chain. Prior to this diff, the selector would treat thrown promises as indicators for when a dependency was resolved _and_ added to a dep cache. Note this assumption is no longer true when user-thrown promises are added to the mix as we can't make guarantees that a user-thrown promise resolving adds anything to the dep cache. The new behavior is that the selector doesn't make the assumption that _all_ thrown promises will resolve and add a value to the dep cache. Instead the selector checks to see if the thrown promises was a Recoil-thrown promise or a user-thrown promise, and based on that decides whether or not to go through dep cache or just force re-evaluation of its dependencies (the latter is the behavior for user-thrown promises). **Note that ultimately means user-thrown promises opt out of some performance optimizations.**

Reviewed By: habond

Differential Revision: D25635015

fbshipit-source-id: a85fd89580bc0886ac98e60270291a0293bf96fd