Configurable cache policies for selectors (#907)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/907

Implements RFC: https://fb.quip.com/rI7AAtzmaaMV

The newest version of Recoil selectors lost the ability to intake custom cache implementation from users due to the special cache requirements of the new selector (it requires a specialized tree-based cache with many constraints built in).

In an effort to hide cache implementation details while providing some way for users to specify the cache eviction and cache equality policies for selectors, we've decided to expose a configurable `CachePolicy` that the user can configure per selector/family.

We'll add more options in the future, but for starters the user will be able to choose from the following policies (types have been simplified for readability):

```
type EqualityPolicy = 'reference' | 'value';
type EvictionPolicy = 'none' | 'lru';

type CachePolicy = {
  equality: EqualityPolicy,
  eviction: EvictionPolicy,
  maxSize?: number, // must be provided when eviction is set to 'lru'
};
```

The default equality policy will be `reference` while the default eviction will be `none`.

Selectors are passed this option in the new `cachePolicy_UNSTABLE` property in the selector options. Families have two new properties: `cachePolicy_UNSTABLE` and `cachePolicyForParams_UNSTABLE`. These are conceptually equivalent to the old properties `cacheImplementation_UNSTABLE` and `cacheImplementationForParams_UNSTABLE`.

### Implementation details (to aid in code review)

The vast majority of files in this diff are just mapping the old `cacheImplementation_UNSTABLE` prop to the new props.

The actual files that contain the core changes are listed here (from high level to low level, so I recommend reviewing in this order):

- `Recoil_CachePolicy.js` - the new cache policy types
- `Recoil_CacheImplementation.js` - defines the types that a Map-based cache should conform to (a simple key-value store). Family params use a Map-based cache for caching params.
- `Recoil_TreeCacheImplementation.js` - defines the types that an opinionated tree-based cache needs to conform to for the purpose of being used in selectors internally.
- `Recoil_cacheFromPolicy.js` - given a cache policy, return a Map-based cache.
- `Recoil_treeCacheFromPolicy.js` - given a cache policy, return a Tree-based cache.
- `Recoil_MapCache.js` - a simple map-based key/value store.
- `Recoil_LRUCache.js` - an LRU cache implementation.
- `Recoil_treeCacheLRU.js` - a helper that uses both `TreeCache` and `LRUCache` to implement a tree cache that has LRU eviction.
- `Recoil_TreeCache.js` - an opinionated tree-based cache.

All the files above have corresponding unit tests.

The rest of the core files in the diff are just changes that use the files and types listed above to expose the cache policy as part of the API:

- `Recoil_UNSTABLE.js`
- `Recoil_evaluatingAtom.js`
- `Recoil_evaluatingAtomFamily.js`
- `Recoil_atomFamily.js`
- `Recoil_constSelector.js`
- `Recoil_errorSelector.js`
- `Recoil_readOnlySelector.js`
- `Recoil_selector.js`
- `Recoil_selectorFamily.js`

Any file not listed in the two sections above is for changing old cache implementation props in userland to the new cache policy props.

Reviewed By: davidmccabe

Differential Revision: D26647188

fbshipit-source-id: 47cb08b66c135b1b1badaf652b8168711bac7c80