Provide "Persist on Read" mode (#1229)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1229

D30691374 allows users to use an atom effect to initialize an atom to a value (potentially based on other atom values) and persist or sync that value on first read.  However, it may not be intuitive that the initialization requires synchronously calling `setSelf()` with an async `Promise`.  This diff adds a new `syncDefault` flag which can be set in the `syncEffect()` hook.  In this mode the atom value will persist on first read from either a synchronous or asynchronous default value or initialized value from an effect.  It will sync the actual value that the atom takes, not just the value based on effect initializations before the `syncEffect()`  (i.e. if a later `syncEffect()` for a different storage in the effects array initializes to a different value that value will be used.

Another ramification of this mode is that it will always try to store a value in the storage when the atom is not set instead of trying to delete the value when the atom is in the default state.

Note that this mode will not persist defaults based on fallback selectors.  That was the gap that still requires use of an atom effect to do the initialization.

Reviewed By: csantos42

Differential Revision: D30718080

fbshipit-source-id: 19b76b81e8db511ac8f792c3f268702396955327