Workaround for missing TreeCache branch
Summary: For users not in the memory management 2020 upgrade, there is a potential atom cleanup issue that subsequent/replaced Recoil Roots will see a reference to an older atom that no longer exists. If that atom was a dependency on a selector, it could cause problems, so now if there's a miss in the branching we reset it and continue forward.

Reviewed By: davidmccabe

Differential Revision: D31431796

fbshipit-source-id: a138221ea287f49d965a343c6bcb73b5d8cfe26e