migrate refine to its own package folder
Summary:
as titled. this should make it possible for us to keep "flow strict" in refine

after this change, the refine source code will NOT be automatically synced to Github. I will need to update our Shipit config after this is committed

files were moved with `hg move` so hopefully don't cause conflicts

Reviewed By: mondaychen

Differential Revision: D31504758

fbshipit-source-id: 33a4f30451412682f4077175f82b6fb9c24e6769