Simplify API by removing Loadables (#1461)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1461

Simplify the API by removing `Loadable`'s and making it more consistent with the selector interface.  Loadables can still be used to capture some corner cases such as setting a synchronous error, but in general can be avoided for simplicity.

The drawback to this approach is more complex handling for reading async items during a write or a compound read, which shouldn't be as common.

Reviewed By: habond

Differential Revision: D32614322

fbshipit-source-id: 668b6333feb7a31c632d9a091738521e26862317