Retain snapshots during asyncMap() (#1555)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1555

When doing an asynchronous snapshot mapping with `Snapshot#asyncMap()` we should retain the mutable snapshot we are operating on during the async map and continue to retain it when returning it to the user.  We can auto-release it after the next tick so it is consistent for the user with synchronous `Snapshot#map()` and they do not need to explicitly `retain()` it if they only use it during the scope of the callback in which they obtained it.

Reviewed By: habond

Differential Revision: D33652974

fbshipit-source-id: a8d53b529b5bd8355a1cdf06bf55682e2a8f008f