Better error reporting / workaround for inconsistent selector values (#1597)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1597

Better detect and clarify error message when selectors do not provide consistent values based on input dependency values.

In development mode this will throw an error.  In production it will log and report to console, then try to reset the selector cache and trundle on.

This workaround is not sufficient to avoid issues and the user should resolve the selector behavior.

Reviewed By: butlersrepos

Differential Revision: D34100576

fbshipit-source-id: 1effa85f19077976b4420222005828f9fdc73a73