Fix test on syncing React and Recoil state changes (#1617)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1617

Fix unit test for synchronizing React and Recoil state changes.  Use `flushSync()` to ensure that state changes are always seen together.

State changes are not guaranteed to be seen atomically by render functions.  Add a new test that confirms that state changes are always seen in the order they are made.

Reviewed By: mondaychen

Differential Revision: D34309556

fbshipit-source-id: 95d5d8c28718a571e4f8251e1b2a61062e1b8042